+++ paywall/src/__tests__/utils/optimisticUnlocking.test.js
@@ -1,10 +1,13 @@
 import * as OptimisticUnlocking from '../../utils/optimisticUnlocking'
+import * as TransactionUtil from '../../utils/getTransaction'
 
 const user = '0xuser'
 const lock = '0xlock'
 const locks = [lock, '0xanother']
 const transaction = '0xtransaction'
 const locksmithUri = 'https://locksmith.unlock-protocol.com'
+
+const { readOnlyProvider } = __ENVIRONMENT_VARIABLES__ /* eslint no-undef: 0 */
 
 const savedTransactions = [
   {
@@ -32,6 +35,7 @@ describe('optimisticUnlocking', () => {
     })
 
     const optimistic = await OptimisticUnlocking.optimisticUnlocking(
+      readOnlyProvider,
       locksmithUri,
       locks,
       user
@@ -43,9 +47,11 @@ describe('optimisticUnlocking', () => {
     expect(OptimisticUnlocking.willUnlock).toHaveBeenCalledTimes(1)
     expect(OptimisticUnlocking.willUnlock).toHaveBeenNthCalledWith(
       1,
+      readOnlyProvider,
       user,
       lock,
-      '0xtransactionHash'
+      '0xtransactionHash',
+      false
     )
   })
 })
@@ -72,10 +78,73 @@ describe('getTransactionsForUserAndLocks', () => {
 })
 
 describe('willUnlock', () => {
-  it('should return true', async () => {
-    expect.assertions(1)
-    expect(await OptimisticUnlocking.willUnlock(user, lock, transaction)).toBe(
-      true
-    )
+  describe('when the transaction does not exist', () => {
+    beforeEach(() => {
+      TransactionUtil.getTransaction = jest.fn(() => {
+        return Promise.resolve(null)
+      })
+    })
+    it('should return true if optimistcIfMissing is true', async () => {
+      expect.assertions(1)
+      const willUnlock = await OptimisticUnlocking.willUnlock(
+        readOnlyProvider,
+        user,
+        lock,
+        transaction,
+        true
+      )
+      expect(willUnlock).toBe(true)
+    })
+    it('should return false if optimistcIfMissing is false', async () => {
+      expect.assertions(1)
+      const willUnlock = await OptimisticUnlocking.willUnlock(
+        readOnlyProvider,
+        user,
+        lock,
+        transaction,
+        false
+      )
+      expect(willUnlock).toBe(false)
+    })
+  })
+  describe('when the transaction exists', () => {
+    it('should return true if the transaction has not been mined', async () => {
+      expect.assertions(1)
+      TransactionUtil.getTransaction = jest.fn(() => {
+        return Promise.resolve({
+          blockNumber: null,
+        })
+      })
+
+      expect(
+        await OptimisticUnlocking.willUnlock(
+          readOnlyProvider,
+          user,
+          lock,
+          transaction,
+          true
+        )
+      ).toBe(true)
+    })
+
+    it('should return false if the transaction has been mined', async () => {
+      expect.assertions(1)
+
+      TransactionUtil.getTransaction = jest.fn(() => {
+        return Promise.resolve({
+          blockNumber: 1337,
+        })
+      })
+
+      expect(
+        await OptimisticUnlocking.willUnlock(
+          readOnlyProvider,
+          user,
+          lock,
+          transaction,
+          true
+        )
+      ).toBe(false)
+    })
   })
 })