+++ paywall/src/paywall-script/index.ts
@@ -1,9 +1,9 @@
 import Postmate from 'postmate'
 import './iframe.css'
 import { setupUnlockProtocolVariable, dispatchEvent } from './utils'
-import { keyExpirationTimestampFor } from '../utils/keyExpirationTimestampFor'
 import { store, retrieve } from '../utils/localStorage'
-import { optimisticUnlocking, willUnlock } from '../utils/optimisticUnlocking'
+import { willUnlock } from '../utils/optimisticUnlocking'
+import { isUnlocked } from '../utils/isUnlocked'
 
 declare let __ENVIRONMENT_VARIABLES__: {
   unlockAppUrl: string
@@ -102,44 +102,22 @@ export class Paywall {
     this.checkKeysAndLock()
   }
 
-  // This function is in charge of assessing if we should unlock the page
-  // First we check the chain, and if any valid key exists, then we unlock
-  // If no valid key exists we check pending transactions to assess them
-  // optimistically.
+  // Will lock or unlock the page based on the current state
   checkKeysAndLock = async () => {
+    const { readOnlyProvider, locksmithUri } = __ENVIRONMENT_VARIABLES__
+    if (!this.userAccountAddress) {
+      return
+    }
     this.lockStatus = undefined
-
-    const { readOnlyProvider } = __ENVIRONMENT_VARIABLES__
-
-    const lockAddresses = Object.keys(this.paywallConfig.locks)
-    const timeStamps = await Promise.all(
-      lockAddresses.map(lockAddress => {
-        return keyExpirationTimestampFor(
-          readOnlyProvider,
-          lockAddress,
-          this.userAccountAddress!
-        )
+    if (
+      await isUnlocked(this.userAccountAddress, this.paywallConfig, {
+        readOnlyProvider,
+        locksmithUri,
       })
-    )
-
-    // If any key is valid, we unlock the page
-    if (timeStamps.some(val => val > new Date().getTime() / 1000)) {
+    ) {
       return this.unlockPage()
     }
-
-    // If not key exists on chain, let's see if we can be optimistic before locking the page.
-    const { locksmithUri } = __ENVIRONMENT_VARIABLES__
-    const optimistic = await optimisticUnlocking(
-      locksmithUri,
-      Object.keys(this.paywallConfig.locks),
-      this.userAccountAddress!
-    )
-    if (optimistic) {
-      return this.unlockPage()
-    }
-
-    // If no key exists, or no transaction exists which could be optimistic, we lock
-    this.lockPage()
+    return this.lockPage()
   }
 
   shakeHands = async () => {
@@ -166,7 +144,14 @@ export class Paywall {
   }
 
   handleTransactionInfoEvent = async ({ hash, lock }: TransactionInfo) => {
-    const optimistic = await willUnlock(this.userAccountAddress!, lock, hash)
+    const { readOnlyProvider } = __ENVIRONMENT_VARIABLES__
+    const optimistic = await willUnlock(
+      readOnlyProvider,
+      this.userAccountAddress!,
+      lock,
+      hash,
+      true // Optimistic if missing
+    )
     if (optimistic) {
       this.unlockPage()
     }