+++ smart-contracts/contracts/mixins/MixinRefunds.sol
@@ -1,21 +1,19 @@
 pragma solidity 0.5.17;
 
-import '@openzeppelin/contracts-ethereum-package/contracts/ownership/Ownable.sol';
 import '@openzeppelin/contracts-ethereum-package/contracts/math/SafeMath.sol';
 import './MixinSignatures.sol';
 import './MixinKeys.sol';
 import './MixinLockCore.sol';
+import './MixinLockManagerRole.sol';
 import './MixinFunds.sol';
-import './MixinApproval.sol';
 
 
 contract MixinRefunds is
-  Ownable,
+  MixinLockManagerRole,
   MixinSignatures,
   MixinFunds,
   MixinLockCore,
-  MixinKeys,
-  MixinApproval
+  MixinKeys
 {
   using SafeMath for uint;
 
@@ -51,9 +49,11 @@ contract MixinRefunds is
    * @dev Invoked by the lock owner to destroy the user's ket and perform a refund and cancellation
    * of the key
    */
-  function expireAndRefundFor(address _keyOwner, uint amount)
-    external
-    onlyOwner
+  function expireAndRefundFor(
+    address _keyOwner,
+    uint amount
+  ) external
+    onlyLockManager
     hasValidKey(_keyOwner)
   {
     _cancelAndRefund(_keyOwner, amount);
@@ -105,9 +105,8 @@ contract MixinRefunds is
   function updateRefundPenalty(
     uint _freeTrialLength,
     uint _refundPenaltyBasisPoints
-  )
-    external
-    onlyOwner
+  ) external
+    onlyLockManager
   {
     emit RefundPenaltyChanged(
       _freeTrialLength,