+++ smart-contracts/test/Lock/disableLock.js
@@ -35,7 +35,10 @@ contract('Lock / disableLock', accounts => {
   })
 
   it('should fail if called by the wrong account', async () => {
-    await reverts(lock.disableLock({ from: keyOwner }), '')
+    await reverts(
+      lock.disableLock({ from: keyOwner }),
+      'MixinLockManager: caller does not have the LockManager role'
+    )
   })
 
   describe('when the lock has been disabled', () => {