+++ unlock-app/src/__tests__/utils/assertNever.test.ts
@@ -1,0 +1,9 @@
+import { assertNever } from '../../utils/assertNever'
+
+describe('assertNever', () => {
+  it('throws', () => {
+    expect.assertions(1)
+
+    expect(assertNever).toThrow()
+  })
+})