+++ unlock-app/src/components/interface/checkout/LockVariations.tsx
@@ -24,7 +24,7 @@ const Checkmark = styled(Svg.Checkmark)`
   background-color: var(--green);
 `
 
-const Ellipsis = styled(Svg.LoadingDots)`
+export const Ellipsis = styled(Svg.LoadingDots)`
   width: 32px;
   margin-left: auto;
   margin-right: 8px;
@@ -197,7 +197,6 @@ export const SoldOutLock = ({
         <LockPrice>{formattedKeyPrice}</LockPrice>
         <ValidityDuration>
           <span>Valid for</span>
-          <br />
           <span>{formattedDuration}</span>
         </ValidityDuration>
         <Arrow />
@@ -222,7 +221,6 @@ export const InsufficientBalanceLock = ({
         <LockPrice>{formattedKeyPrice}</LockPrice>
         <ValidityDuration>
           <span>Valid for</span>
-          <br />
           <span>{formattedDuration}</span>
         </ValidityDuration>
         <Arrow />
@@ -248,7 +246,6 @@ export const DisabledLock = ({
         <LockPrice>{formattedKeyPrice}</LockPrice>
         <ValidityDuration>
           <span>Valid for</span>
-          <br />
           <span>{formattedDuration}</span>
         </ValidityDuration>
         <Arrow />
@@ -274,7 +271,6 @@ export const PurchaseableLock = ({
         <LockPrice>{formattedKeyPrice}</LockPrice>
         <ValidityDuration>
           <span>Valid for</span>
-          <br />
           <span>{formattedDuration}</span>
         </ValidityDuration>
         <Arrow />
@@ -298,7 +294,6 @@ export const ProcessingLock = ({
         <LockPrice>{formattedKeyPrice}</LockPrice>
         <ValidityDuration>
           <span>Valid for</span>
-          <br />
           <span>{formattedDuration}</span>
         </ValidityDuration>
         <Ellipsis />
@@ -323,7 +318,6 @@ export const ConfirmedLock = ({
         <LockPrice>{formattedKeyPrice}</LockPrice>
         <ValidityDuration>
           <span>Valid for</span>
-          <br />
           <span>{formattedDuration}</span>
         </ValidityDuration>
         <Checkmark />