+++ unlock-js/src/v12/index.js
@@ -4,6 +4,8 @@ import getLock from './getLock'
 import purchaseKey from './purchaseKey'
 import updateKeyPrice from './updateKeyPrice'
 import withdrawFromLock from './withdrawFromLock'
+import configureUnlock from './configureUnlock'
+import isLockManager from './isLockManager'
 
 export default {
   version: 'v12',
@@ -14,4 +16,6 @@ export default {
   updateKeyPrice,
   purchaseKey,
   withdrawFromLock,
+  configureUnlock,
+  isLockManager,
 }