+++ unlock-js/src/v13/index.js
@@ -5,6 +5,8 @@ import purchaseKey from './purchaseKey'
 import updateKeyPrice from './updateKeyPrice'
 import withdrawFromLock from './withdrawFromLock'
 import initializeTemplate from './initializeTemplate'
+import configureUnlock from './configureUnlock'
+import isLockManager from './isLockManager'
 
 export default {
   version: 'v13',
@@ -16,4 +18,6 @@ export default {
   updateKeyPrice,
   purchaseKey,
   withdrawFromLock,
+  configureUnlock,
+  isLockManager,
 }